Move smearer and LUT to dedicated file
